Java Code Convention 1
