+++ src/main/java/mypackage/minuyim/baseball/Main.java
@@ -3,12 +3,11 @@ package mypackage.minuyim.baseball;
 import mypackage.minuyim.baseball.MainGame;
 
 public class Main {
+    public static void main(String[] args) {
+        MainGame game = new MainGame();
+        game.start();
 
-	public static void main(String[] args) {
-		MainGame game = new MainGame();
-		game.start();
-		
-		return;
-	}
+        return;
+    }
 
 }
