--- src/main/java/mypackage/minuyim/baseball/util/PrintHint.java
@@ -1,15 +1,0 @@
-package mypackage.minuyim.baseball.util;
-
-public class PrintHint {
-	public void printHint(int strike, int ball) {
-		if (strike == 0 && ball == 0) {
-			System.out.println("낫싱");
-		} else if (ball == 0) {
-			System.out.println(strike + "스트라이크");
-		} else if (strike == 0) {
-			System.out.println(ball + "볼");
-		} else {
-			System.out.println(strike + "스트라이크 " + ball + "볼");
-		}
-	}
-}
